local-ai-backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
black==23.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.121.0
filelock==3.19.1
flake8==6.1.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mpmath==1.3.0
mypy==1.7.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
onnxruntime-directml==1.23.0
opencv-python==4.12.0.88
openvino==2025.3.0
openvino-telemetry==2025.2.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
polars==1.35.1
polars-runtime-32==1.35.1
protobuf==6.33.0
psutil==7.1.3
pycodestyle==2.11.1
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.1.0
pyparsing==3.2.5
pyreadline3==3.5.4
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
segment-anything==1.0
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
torch==2.7.1
torchvision==0.22.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.226
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
